requirements.txt
absl-py==0.13.0
astor==0.8.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
decorator==4.4.2
gast==0.2.2
glob2==0.7
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.5.0
ISR==2.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
moviepy==1.0.3
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
Pillow==8.2.0
proglog==0.1.9
protobuf==3.17.3
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
